Trimming files and fixing problems
I had an issue where I believed the chi square analysis was no longer working, but it was an issue between two versions of HPAStainR and the naming system of the arguement "cancer_analysis", which went from c(TRUE, FALSE) to c("Normal, "Cancer", "Both") to c("both", "normal", "cancer"). I settled of the final version because all other string arguements are lowercase.